implemented d3 tree graph with react, added to webpack
