+++ package.json
@@ -37,7 +37,9 @@
     "html-webpack-plugin": "^4.5.1",
     "mini-css-extract-plugin": "^1.3.5",
     "nodemon": "^2.0.7",
+    "resize-observer-polyfill": "^1.5.1",
     "style-loader": "^2.0.0",
+    "use-resize-observer": "^7.0.0",
     "webpack": "^5.20.0",
     "webpack-cli": "^4.5.0",
     "webpack-dev-server": "^3.11.2"
@@ -46,6 +48,7 @@
     "ace-builds": "^1.4.12",
     "codemirror": "^5.59.2",
     "cors": "^2.8.5",
+    "d3": "^6.5.0",
     "electron-builder": "^22.9.1",
     "electron-fetch": "^1.7.3",
     "electron-is-dev": "^1.2.0",
@@ -67,6 +70,8 @@
     "react-dom": "^17.0.1",
     "react-router-dom": "^5.2.0",
     "react-tabs": "^3.2.0",
+    "react-tree-graph": "^6.0.0",
+    "svg-url-loader": "^7.1.1",
     "wait-on": "^5.2.1"
   }
 }
