+++ src/Components/useResizeObserver.js
@@ -1,0 +1,21 @@
+// import { useEffect, useState } from "react";
+// import ResizeObserver from "resize-observer-polyfill";
+
+// const useResizeObserver = ref => {
+//   const [dimensions, setDimensions] = useState(null);
+//   useEffect(() => {
+//     const observeTarget = ref.current;
+//     const resizeObserver = new ResizeObserver(entries => {
+//       entries.forEach(entry => {
+//         setDimensions(entry.contentRect);
+//       });
+//     });
+//     resizeObserver.observe(observeTarget);
+//     return () => {
+//       resizeObserver.unobserve(observeTarget);
+//     };
+//   }, [ref]);
+//   return dimensions;
+// };
+
+// export default useResizeObserver;
