+++ src/public/index.css
@@ -11,6 +11,7 @@
 
 body {
   background-color: #111212;
+  /* background-color: white; */
   font-family: monospace;
 }
 
@@ -32,6 +33,9 @@ body {
  text-align: center;
  }
 
+ .node__root {
+   color: red;
+ }
 
 
 /* .URISubmitButton {
@@ -134,7 +138,7 @@ body {
   grid-template-rows: 1fr;
   gap: 0px 0px;
   grid-template-areas:
-    "dropDown . codeboxContainer";
+    "dropDown tree codeboxContainer";
 }
 
 .dropDown {
@@ -183,4 +187,38 @@ body {
   border-color: #aaa;
   color: black;
   border-radius: 5px 5px 0 0;
+}
+
+.circle {
+  background-color: black;
+}
+.node {
+  background-color: black;
+}
+
+div.custom-container {
+	background-color: #242424;
+}
+
+svg.custom .node circle {
+	fill: #f4f4f4;
+	stroke: #2593B8;
+	stroke-width: 2px;
+}
+
+svg.custom .node text {
+	font-size: 16px;
+	background-color: #444;
+	fill: #f4f4f4;
+	text-shadow: 0 1px 4px black;
+}
+
+svg.custom .node {
+	cursor: pointer;
+}
+
+svg.custom path.link {
+	fill: none;
+	stroke: #32a8cf;
+	stroke-width: 2px;
 }
